dom/imptests/webapps/DOMCore/tests/submissions/Ms2ger/test_Document-createComment.html

<!DOCTYPE html>
<title>Document.createComment</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
test(function() {
  var c = document.createComment("a -- b");
  assert_true(c instanceof Comment);
  assert_true(c instanceof Node);
  assert_equals(c.ownerDocument, document);
  assert_equals(c.data, "a -- b");
  assert_equals(c.nodeValue, "a -- b");
  assert_equals(c.textContent, "a -- b");
  assert_equals(c.length, 6);
  assert_equals(c.nodeType, 8);
  assert_equals(c.nodeName, "#comment");
  assert_equals(c.hasChildNodes(), false);
  assert_equals(c.childNodes.length, 0);
  assert_equals(c.firstChild, null);
  assert_equals(c.lastChild, null);
});
</script>
